+++ src/components/LogOutButton.tsx
@@ -1,7 +1,7 @@
-"use client";
+'use client';
 
 import { Button } from '@mui/material';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 function Logout() {
 	const router = useRouter();
