+++ src/components/ResetButton.tsx
@@ -1,8 +1,8 @@
-"use client";
+'use client';
 
 import WarningIcon from '@mui/icons-material/Warning';
 import { Box, Button, useTheme } from '@mui/material';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import CustomTooltip from './CustomTooltip';
 
 function ResetButton() {
